splitt modules

+++ components/asic/include/frequency_transition_bmXX.h
@@ -13,7 +13,7 @@ extern const char *FREQUENCY_TRANSITION_TAG;
  * 
  * @param frequency The frequency to set in MHz
  */
-typedef void (*set_hash_frequency_fn)(float frequency);
+typedef void (*set_hash_frequency_fn)(double frequency);
 
 /**
  * @brief Transition the ASIC frequency to a target value
@@ -26,6 +26,6 @@ typedef void (*set_hash_frequency_fn)(float frequency);
  * @param asic_type The type of ASIC chip (for logging purposes only)
  * @return bool True if the transition was successful, false otherwise
  */
-bool do_frequency_transition(float target_frequency, set_hash_frequency_fn set_frequency_fn, int asic_type);
+bool do_frequency_transition(double target_frequency, set_hash_frequency_fn set_frequency_fn, int asic_type);
 
 #endif // FREQUENCY_TRANSITION_H
